+++ src/app/(no-bottom-menu)/@modal/nearby/[id]/page.tsx
@@ -1,0 +1,19 @@
+"use client";
+
+import { useState } from "react";
+import { KakaoLoginDialog } from "@/components/modal";
+import { useRouter } from "next/navigation";
+
+export default function LoginModal() {
+  const router = useRouter();
+  const [isOpen, setIsOpen] = useState(true);
+  return (
+    <KakaoLoginDialog
+      isOpen={isOpen}
+      onClose={() => {
+        setIsOpen(false);
+        router.back();
+      }}
+    />
+  );
+}
