+++ src/app/(no-bottom-menu)/write/page.tsx
