+++ src/app/(with-bottom-menu)/nearby/page.tsx
@@ -7,7 +7,9 @@ import NearbyFallback from "@/components/fallback/nearby-fallback";
 export default function Page() {
   return (
     <Suspense fallback={<NearbyFallback />}>
-      <LetterContent />
+      <main className="flex-1 min-h-0 overflow-y-auto pb-20 [&>article]:min-h-full">
+        <LetterContent />
+      </main>
     </Suspense>
   );
 }
