+++ src/app/(with-bottom-menu)/settings/page.tsx
@@ -1,0 +1,7 @@
+export default function Page() {
+  return (
+    <div>
+      <h1>Setting Pg</h1>
+    </div>
+  );
+}
