+++ src/components/navigation/bottom-navigation.tsx
@@ -1,39 +1,21 @@
 "use client";
 
 import { NAV_ITEMS } from "@/constants/nav-items";
-import { usePathname, useRouter } from "next/navigation";
-import { useState } from "react";
-import { KakaoLoginDialog } from "../modal";
+import { usePathname } from "next/navigation";
+import Link from "next/link";
 
-interface BottomNavigationProps {
-  isLoggedIn: boolean;
-}
-
-export default function BottomNavigation({ isLoggedIn }: BottomNavigationProps) {
+export default function BottomNavigation() {
   const pathname = usePathname();
-  const router = useRouter();
-  const [isLoginModalOpen, setIsLoginModalOpen] = useState(false);
-
-  const PROTECTED_ROUTES = ["/write", "/settings"];
-
-  const handleNavigation = (path: string) => {
-    if (PROTECTED_ROUTES.includes(path) && !isLoggedIn) {
-      setIsLoginModalOpen(true);
-      return;
-    }
-
-    router.push(path);
-  };
 
   return (
     <>
       <nav id="bottom-navigation" className="fixed bottom-0 z-10 flex w-full flex-col items-center justify-center">
         <div className="absolute bottom-0 w-full bg-white rounded-t-2xl shadow p-2 flex items-center justify-around z-50 border-t border-secondary">
           {NAV_ITEMS.map((item) => (
-            <div
+            <Link
+              href={item.path}
               key={item.label}
-              onClick={() => handleNavigation(item.path)}
-              className={`flex flex-col items-center mb-1 cursor-pointer ${
+              className={`flex flex-col items-center mb-1 ${
                 pathname === item.path
                   ? "text-brand-primary font-bold"
                   : "text-text-secondary/50 hover:text-brand-primary"
@@ -41,12 +23,10 @@ export default function BottomNavigation({ isLoggedIn }: BottomNavigationProps) 
             >
               {item.icon}
               <span className="text-[11px]">{item.label}</span>
-            </div>
+            </Link>
           ))}
         </div>
       </nav>
-
-      <KakaoLoginDialog isOpen={isLoginModalOpen} onClose={() => setIsLoginModalOpen(false)} />
     </>
   );
 }
