+++ src/hooks/auth/index.ts
@@ -1,0 +1,1 @@
+export { default as useAuth } from "./use-auth";
