--- src/providers/session-provider.tsx
@@ -1,14 +1,0 @@
-"use client";
-
-import { Session } from "next-auth";
-import { SessionProvider } from "next-auth/react";
-import { ReactNode } from "react";
-
-type Props = {
-  session?: Session | null;
-  children: ReactNode;
-};
-
-export default function AuthSessionProvider({ session, children }: Props) {
-  return <SessionProvider session={session}>{children}</SessionProvider>;
-}
